sks ta: utils to create serialized object attributes
Add functions to create, fill, modify a serialized object.
This relies on a internal structure holding the serial buffer
reference and functions that play with realloc and free routines
to add/remove attribute item from the serialized attributes list
of an object.

A generated serialized object is a memory buffer starting with
a header structure followed by each attribute content (id, size
and value).

Add a debug function to dump the content of a serialized object.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>