sks ta: find objects
3 invocation commands: SKS_CMD_FIND_OBJECTS*. A command initializes
a search and establishes list of matching objects. A command allows
client to get handles (by burst) for matching objects. Last command
finalize the search and releases the search resources. When the
session closes (whatever reason) the search resource must be released.

The object search resources are referenced by a "find object" context
structure. It stores all allocated resources: the target matching
attributes list, a list of object handles for the client and a list
of persistent objects that matched the search but have not yet registered
through a handle within the session.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>