libsks: SKS_CK_VENDOR_INVALID_ID reflects CK_VENDOR_INVALID_ID
SKS_CK_VENDOR_INVALID_ID denotes an identifier value that is invalid.
It may be used in arguments when no specific value was provided for
an argument entry by the client. It is a vendor specific ID SKS TA
creates in the PKCS#11 implementation, under macro CK_VENDOR_INVALID_ID.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>