sks ta: fix object database read at TA creation
In the previous implementation, the persistent objects UUID
was referred in the buffer allocate to store the persistent
object database instance in TA RAM. At object creation/destruction,
the database is reallocated hence UUID references are not more
valid. This change allocates a buffer to store the object UUID
when the instance of the object in created at TA initialization.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>