sks ta: client template may not define class and type
It is not mandatory for a client template to define the
object class and type-in-class. For example, templates
used in Find-Object function may not. However, created
object must define consistent class and type-in-class.

As a consequence, get_bool() should return a value when
the boolean attribute is not define: returns value FALSE.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>