sks ta: generate key pair &  asymm cipher/authen
This is a big change the merges few changes:
- introduce get_u32_attribute() to easy read of 32bit value
  attribute in an object, mainly for identifiers, flag masks
  and size fields.
- introduce RSA and EC private and public object import.
  EC object import will not create a valid key as EC public
  point coordinate require DER format support.
- introduce RSA and EC private and public key generation.
- store EC curve parameters in DER format. The numbers were
  generated with openssl tools for the standard NIST and
  Brainpool curves, by name and by explicit parameters.
- Support some RSA PSS, some RSA OAEP,  pre-hashed ECDSA,

The active processing stores the TEE type of the loaded
key. This is required for EC processing: PKCS#11 does not
differentiate ECDSA keys from ECDH keys whereas GPD TEE does.
Therefore, if a PKCS#11 EC key was used for ECDSA and needs
to be used for ECDH, the TEE key shall be reloaded from the
key attributes.

    Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>