libsks: link against libteec
Previously libsks was not linked dynamically against libteec.
This was not noticed because the xtest application is linked against both
libraries and the symbols could be resolved during runtime.

However loading libsks as a standalone pkcs#11 library failed because of the
missing dynamic linking information.
Add libteec as a dynamic link dependency and depend the build-libsks target on
the build-libteec target.

Signed-off-by: Rouven Czerwinski <r.czerwinski@pengutronix.de>