[toup] libsks: fix rv when buffer too small is expected
From the specification:

If pBuf is NULL_PTR, then all that the function does is return (in
*pulBufLen) a number of bytes which would suffice to hold the
cryptographic output produced from the input to the function. This
number may somewhat exceed the precise number of bytes needed, but
should not exceed it by a large amount. CKR_OK is returned by the
function.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>